sessly-b/requirements.txt
asgiref==3.11.0
black==25.12.0
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dj-database-url==2.2.0
Django==6.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.28.1
google-api-python-client==2.149.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.72.0
gunicorn==23.0.0
httplib2==0.31.0
idna==3.11
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
proto-plus==1.26.1
protobuf==6.33.2
psycopg==3.2.12
psycopg-binary==3.2.12
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dotenv==1.0.1
pytokens==0.3.0
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.4
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.0
whitenoise==6.7.0
